refactor(runner): extract shared createAsyncLocalStore helper
The write-debug-context store (runtime.ts) and the action-execution
window (builder/action-context.ts) each carried a near-identical
`{ getStore, run }` store: Deno `AsyncLocalStorage` when available, a
promise-aware synchronous fallback otherwise. Hoist the one shape into
`@commonfabric/utils/async-local-store` (`createAsyncLocalStore<T>()`)
and import it at both call sites. Fallback semantics are preserved
verbatim — previous value restored in a `.finally` for promise results,
synchronously otherwise.
