+++ packages/runner/src/runtime.ts
@@ -77,6 +77,7 @@ import { ExtendedStorageTransaction } from "./storage/extended-storage-transacti
 import { isCellScope, normalizeCellScope } from "./scope.ts";
 import { toURI } from "./uri-utils.ts";
 import { isDeno } from "@commonfabric/utils/env";
+import { createAsyncLocalStore } from "@commonfabric/utils/async-local-store";
 import { popFrame, pushFrame } from "./builder/pattern.ts";
 import type { Frame } from "./builder/types.ts";
 import type { ConsoleMessage } from "./interface.ts";
@@ -113,42 +114,6 @@ const isFullNormalizedLinkShape = (
     Array.isArray(link.path) &&
     (link.scope === undefined || isCellScope(link.scope));
 };
-
-interface WriteDebugContextStore<T> {
-  getStore(): T | undefined;
-  run<R>(value: T, fn: () => R): R;
-}
-
-class FallbackAsyncLocalStorage<T> implements WriteDebugContextStore<T> {
-  #store: T | undefined;
-
-  getStore(): T | undefined {
-    return this.#store;
-  }
-
-  run<R>(value: T, fn: () => R): R {
-    const previous = this.#store;
-    this.#store = value;
-    try {
-      const result = fn();
-      if (result instanceof Promise) {
-        return result.finally(() => {
-          this.#store = previous;
-        }) as R;
-      }
-      this.#store = previous;
-      return result;
-    } catch (error) {
-      this.#store = previous;
-      throw error;
-    }
-  }
-}
-
-const WriteDebugContextStorage = isDeno()
-  ? (await import("node:async_hooks"))
-    .AsyncLocalStorage as new <T>() => WriteDebugContextStore<T>
-  : FallbackAsyncLocalStorage;
 
 // @ts-ignore - This is temporary to debug integration test
 Error.stackTraceLimit = 500;
@@ -346,7 +311,7 @@ export class Runtime {
   private readonly spaceNameToDid = new Map<string, MemorySpace>();
   private defaultFrame?: Frame;
   private queues = new Map<string, AsyncSemaphoreQueue>();
-  private writeDebugContext = new WriteDebugContextStorage<string>();
+  private writeDebugContext = createAsyncLocalStore<string>();
   private cfcStats: CfcRuntimeStats = initialCfcRuntimeStats();
 
   constructor(options: RuntimeOptions) {
